+++ database/migrations/2020_08_03_145357_create_replies_table.php
@@ -3,13 +3,13 @@
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateRepliesTable extends Migration 
+class CreateRepliesTable extends Migration
 {
 	public function up()
 	{
 		Schema::create('replies', function(Blueprint $table) {
-            $table->increments('id');
-            $table->integer('topic_id')->unsigned()->default(0)->index();
+            $table->bigIncrements('id');
+            $table->bigInteger('topic_id')->unsigned()->default(0)->index();
             $table->bigInteger('user_id')->unsigned()->default(0)->index();
             $table->text('content');
             $table->timestamps();
